Move parse_csv files to validator dir
