+++ usecase/validator/parse_csv_cmd_handler.go
@@ -1,4 +1,4 @@
-package indexing
+package validator
 
 import (
 	"context"
